pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/icarus"]

[project]
name = "icarus-cds"
version = "0.3.0"
authors = [{ name = "Cody", email = "cody@dkdc.dev" }]
description = "Python composable data stack solution accelerator"
readme = "readme.md"
requires-python = ">=3.9"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    # secret management
    'python-dotenv',
    # cli
    'typer',
    # cloud
    'gcsfs',
    # data
    'ibis-framework[duckdb,polars,deltalake,examples]',
    'Faker',
    # visualization
    'plotly',
    'great-tables',
    # dashboards and apps
    'shiny>=1.0.0',
    'shinywidgets>=0.3.3',
    'shinyswatch>=0.7.0',
]

[project.urls]
"Homepage" = "https://github.com/lostmygithubaccount/Icarus"
"Bug Tracker" = "https://github.com/lostmygithubaccount/Icarus/issues"

[project.scripts]
icarus = "icarus.cli:app"

[tool.ruff]
extend-include = ["*.ipynb"]
